Keyboard fix
